requirements.txt
alice-blue==1.0.1
attrs==20.3.0
autobahn==19.11.2
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
enum34==1.1.10
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-auth==1.32.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kiteconnect==3.9.2
oauthlib==3.1.1
packaging==20.9
pip==21.0.1
protlib==1.4
protobuf==3.17.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7.2
service-identity==18.1.0
setuptools==54.1.2
six==1.15.0
soupsieve==2.2.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
websocket-client==1.1.0
zope.interface==5.3.0
